+++ src/controller.ts
@@ -1,6 +1,6 @@
 import { Application } from "pixi.js";
-import { Minigame, MinigameDelegate } from "./minigame";
-import { KeyboardMinigame } from "./keyboardMinigame";
+import { Minigame, MinigameDelegate } from "./minigames/minigame";
+import { KeyboardMinigame } from "./minigames/keyboardMinigame";
 
 /** Controls the flow of the game. */
 export class Controller implements MinigameDelegate {
@@ -8,16 +8,16 @@ export class Controller implements MinigameDelegate {
 
   constructor(private readonly app: Application) { }
 
-  finished(passed: boolean) {
+  start() {
+    this.currentMinigame = new KeyboardMinigame(this.app, this);
+    this.currentMinigame.attach();
+  }
+
+  onMinigameEnd(passed: boolean) {
     this.currentMinigame!.detach();
     this.currentMinigame = undefined;
     console.log('Player', passed ? 'won' : 'lost');
     // For now just restart the same minigame again.
     this.start();
-  }
-
-  start() {
-    this.currentMinigame = new KeyboardMinigame(this.app, this);
-    this.currentMinigame.attach();
   }
 }
