+++ src/index.ts
@@ -18,6 +18,5 @@ class Game {
   }
 }
 
-
 const game = new Game();
 game.start();
